# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nifty100 Data Foundation API",
    "description": "Internal API over the Nifty100 fundamentals/screener/analytics data foundation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/peers/groups": {
      "get": {
        "tags": [
          "peers"
        ],
        "summary": "List Peer Groups",
        "description": "List peer groups.",
        "operationId": "list_peer_groups_api_v1_peers_groups_get",
        "parameters": [
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/peers/groups/{peer_group_name}": {
      "get": {
        "tags": [
          "peers"
        ],
        "summary": "Peer Group Members",
        "description": "Peer group members.",
        "operationId": "peer_group_members_api_v1_peers_groups__peer_group_name__get",
        "parameters": [
          {
            "name": "peer_group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Peer Group Name"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/peers/{company_id}/percentiles": {
      "get": {
        "tags": [
          "peers"
        ],
        "summary": "Company Peer Percentiles",
        "description": "Company peer percentiles.",
        "operationId": "company_peer_percentiles_api_v1_peers__company_id__percentiles_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/peers/{group_name}": {
      "get": {
        "tags": [
          "peers"
        ],
        "summary": "Peer Group Detail",
        "description": "Peer group detail.",
        "operationId": "peer_group_detail_api_v1_peers__group_name__get",
        "parameters": [
          {
            "name": "group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Name"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/valuation/{company_id}": {
      "get": {
        "tags": [
          "valuation"
        ],
        "summary": "Valuation History",
        "description": "Valuation history.",
        "operationId": "valuation_history_api_v1_valuation__company_id__get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/valuation/{company_id}/latest": {
      "get": {
        "tags": [
          "valuation"
        ],
        "summary": "Valuation Latest",
        "description": "Valuation latest.",
        "operationId": "valuation_latest_api_v1_valuation__company_id__latest_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}